+++ admin/src/containers/Admin/Main.tsx
@@ -1,10 +1,7 @@
 import {
   Component,
-  createEffect,
   createSignal,
-  For,
   Match,
-  on,
   onCleanup,
   onMount,
   Show,
@@ -17,36 +14,45 @@ import Creator from "./modules/Creator";
 import { IPost } from "~/api/types";
 
 import "~/styles/admin.scss";
-import SidePanel from "./modules/PostPanel";
+import PostPanel from "./modules/PostPanel";
 import { createStore } from "solid-js/store";
 import Navigator from "./modules/Navigator";
+import Metrics from "./metrics/Metrics";
 
 const MODEWIDTH = 750;
 
 const Main: Component = () => {
   const [posts, setPosts] = createSignal<IPost[]>([]);
-  const [showCreator, setShowCreator] = createSignal<boolean>(false);
-  const [selectedPost, setSelectedPost] = createSignal<IPost>();
 
   const [rect, setRect] = createSignal({
     height: window.innerHeight,
     width: window.innerWidth,
   });
 
-  const [adminState, setAdminStore] = createStore<{
-    editor: boolean;
+  const [adminState, setAdminState] = createStore<{
     posts: boolean;
-  }>({ editor: false, posts: true });
+    metrics: boolean;
+    editor: boolean;
+    editorContent: {
+      creator: boolean;
+      editorPost: IPost;
+    };
+  }>({
+    posts: true,
+    metrics: false,
+    editor: false,
+    editorContent: { creator: false, editorPost: null },
+  });
 
   const handler = (_: Event) => {
     setRect({ height: window.innerHeight, width: window.innerWidth });
     if (window.innerWidth <= MODEWIDTH) {
       if (adminState.editor) {
-        setAdminStore({ posts: false });
+        setAdminState({ posts: false });
       }
     }
     if (window.innerWidth >= MODEWIDTH) {
-      setAdminStore({ posts: true });
+      setAdminState({ posts: true });
     }
   };
 
@@ -67,23 +73,29 @@ const Main: Component = () => {
   };
 
   const editorUpdatePostSelector = (post: IPost) => () => {
-    setShowCreator(false);
-    setSelectedPost(post);
-    setAdminStore({ editor: true });
+    setAdminState({
+      editor: true,
+      posts: true,
+      metrics: false,
+      editorContent: { creator: false, editorPost: post },
+    });
     if (window.innerWidth <= MODEWIDTH) {
       if (adminState.editor) {
-        setAdminStore({ posts: false });
+        setAdminState({ posts: false });
       }
     }
   };
 
   const editorCreateSelector = () => {
-    setSelectedPost();
-    setShowCreator(true);
-    setAdminStore({ editor: true });
+    setAdminState({
+      editor: true,
+      posts: true,
+      metrics: false,
+      editorContent: { creator: true, editorPost: null },
+    });
     if (window.innerWidth <= MODEWIDTH) {
       if (adminState.editor) {
-        setAdminStore({ posts: false });
+        setAdminState({ posts: false });
       }
     }
   };
@@ -95,37 +107,48 @@ const Main: Component = () => {
           <div class="admin-navigator">
             <Navigator
               editorCreateSelector={editorCreateSelector}
-              setAdminStore={setAdminStore}
-              setSelectedPost={setSelectedPost}
+              setAdminState={setAdminState}
             />
           </div>
-          <div
-            class={`admin-panel ${
-              rect().width >= MODEWIDTH && adminState.editor && adminState.posts
-                ? "multi-mode"
-                : "single-mode"
-            }`}
-          >
-            <Show when={adminState.posts}>
-              <SidePanel
-                posts={posts()}
-                editorUpdatePostSelector={editorUpdatePostSelector}
-                selectedPost={selectedPost}
-              />
-            </Show>
-            <Show when={adminState.editor}>
-              <div class="admin-panel-editor">
-                <Switch>
-                  <Match when={selectedPost()}>
-                    <Updater post={selectedPost} fetchAll={getAllPosts} />
-                  </Match>
-                  <Match when={showCreator()}>
-                    <Creator />
-                  </Match>
-                </Switch>
+          <Switch>
+            <Match when={!adminState.metrics}>
+              <div
+                class={`admin-panel ${
+                  rect().width >= MODEWIDTH &&
+                  adminState.editor &&
+                  adminState.posts
+                    ? "multi-mode"
+                    : "single-mode"
+                }`}
+              >
+                <Show when={adminState.posts}>
+                  <PostPanel
+                    posts={posts()}
+                    adminState={adminState}
+                    editorUpdatePostSelector={editorUpdatePostSelector}
+                  />
+                </Show>
+                <Show when={adminState.editor}>
+                  <div class="admin-panel-editor">
+                    <Switch>
+                      <Match when={adminState.editorContent.editorPost}>
+                        <Updater
+                          adminState={adminState}
+                          fetchAll={getAllPosts}
+                        />
+                      </Match>
+                      <Match when={adminState.editorContent.creator}>
+                        <Creator />
+                      </Match>
+                    </Switch>
+                  </div>
+                </Show>
               </div>
-            </Show>
-          </div>
+            </Match>
+            <Match when={adminState.metrics}>
+              <Metrics></Metrics>
+            </Match>
+          </Switch>
         </div>
       </div>
     </main>