+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.6] - 2023-03-30
+
+#### Changed
+
+- admin console now is using a single store for state
+- lowered the pixel width and height of the admin page svgs
+
+#### Added
+
+- started state values for the admin metrics page
+
 ## [0.8.5] - 2023-03-30
 
 #### Added